Cleaned up construct solution and added output directory for submissions
