+++ submissions/byproducts/note.txt
@@ -1,0 +1,1 @@
+Herein lines the intermediate representations of the construct ordering
