+++ submissions/note.txt
@@ -1,0 +1,1 @@
+Ah a note, such an elegant way to commit an empty directory. Herein lies the graveyarrd of submissions which will probably never get checked in.
